+++ src/main/java/com/offtomarket/mod/OffToMarket.java
@@ -14,11 +14,14 @@ import com.offtomarket.mod.registry.ModBlockEntities;
 import com.offtomarket.mod.registry.ModBlocks;
 import com.offtomarket.mod.registry.ModItems;
 import com.offtomarket.mod.registry.ModMenuTypes;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.RegisterCommandsEvent;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.MissingMappingsEvent;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig.Type;
@@ -82,5 +85,27 @@ public class OffToMarket {
     @SubscribeEvent
     public void onRegisterCommands(RegisterCommandsEvent event) {
         DebugCommands.register(event.getDispatcher());
+    }
+
+    @SubscribeEvent
+    public void onMissingMappings(MissingMappingsEvent event) {
+        // Remap old "trading_bin" IDs to "trading_ledger" for world compatibility
+        ResourceLocation oldId = new ResourceLocation(MODID, "trading_bin");
+
+        for (var mapping : event.getMappings(ForgeRegistries.Keys.BLOCKS, MODID)) {
+            if (mapping.getKey().equals(oldId)) {
+                mapping.remap(ModBlocks.TRADING_LEDGER.get());
+            }
+        }
+        for (var mapping : event.getMappings(ForgeRegistries.Keys.ITEMS, MODID)) {
+            if (mapping.getKey().equals(oldId)) {
+                mapping.remap(ModItems.TRADING_LEDGER_ITEM.get());
+            }
+        }
+        for (var mapping : event.getMappings(ForgeRegistries.Keys.BLOCK_ENTITY_TYPES, MODID)) {
+            if (mapping.getKey().equals(oldId)) {
+                mapping.remap(ModBlockEntities.TRADING_LEDGER.get());
+            }
+        }
     }
 }